Support proofs with more than 19 different statements
This required changing the way we decompress proofs: instead of going
letter-by-letter, we need to support UUA=120 and such.

Coincidentally, the very first proof that uses those statements is also
the first to use a floating hypothesis that isn't required by the
theorem statement (it uses "wff ch", even though the theorem is
expressed only in terms of phi and psi, because some intermediate
results need chi). The commit also fixes a bug when that happens.